Moved scene into its own package
